Start pulling real fixtures for data.
