+++ lib/dpul_collections_web/live/search_live/search_state.ex
@@ -8,7 +8,8 @@ defmodule DpulCollectionsWeb.SearchLive.SearchState do
       sort_by: valid_sort_by(params),
       page: (params["page"] || "1") |> String.to_integer(),
       per_page: (params["per_page"] || "50") |> String.to_integer(),
-      filter: params["filter"] || %{}
+      filter: params["filter"] || %{},
+      extra_params: %{}
     }
   end
